requirements.txt
docx2txt==0.7
numpy==1.15.4
opencv-contrib-python==3.4.4.19
Pillow==5.3.0
pytesseract==0.2.5
Wand==0.4.5
xlrd==1.1.0